+++ inst/tinytest/test-hypotheses.R
@@ -207,4 +207,9 @@ h <- hypotheses(fm1,
                  joint = c("SexFemale", "age"))
 
 
+# Issue #1102: hypotheses() should not be called twice on the same object
+mod <- lm(mpg ~ hp + wt + factor(cyl), data = mtcars)
+mod <- hypotheses(mod)
+expect_error(hypotheses(mod), pattern = "twice on the same object")
+
 rm(list = ls())