+++ NEWS.md
@@ -9,6 +9,7 @@ New function:
 Minor:
 
 * Do not raise extraneous warning for `survey` package models when the `by` argument is not used.
+* Informative error when `hypotheses()` is called twice on the same object.
 
 Bugs: